app/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-core
langchain-google-vertexai
langchain-chroma
langchain-community
google-cloud-secret-manager
google-cloud-logging
google-auth
google-cloud-storage
firebase-admin
chroma
pypdf
fpdf
youtube-transcript-api
pytubecd